backend/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
mysql-connector-python==8.1.0
python-dotenv==1.0.0
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
Werkzeug==2.3.6
Flask-SocketIO==5.3.4
python-socketio==5.8.0
eventlet==0.33.3